MAJOR change: jump_gravity_acceleration
There used to be a jump_gravity_max varaiale, which was supposed to achieve a lower gravity when the player is jumping up. I realized my logic was flawed, so I renamed the varaiable to jump_gravity_acceleration and fixed the logic.

I also rebalanced the varaiables to be less arbitrary.